Configure travis-ci to also build on ubuntu trusty
Now we have Ubuntu 12.04 and 14.04 automatic test builds.

I had to remove ffmpeg from the package list; this was actually just
installing an empty redirection package on 12.04, that empty package
was removed on 14.04, see:

http://askubuntu.com/questions/432542/is-ffmpeg-missing-from-the-official-repositories-in-14-04

(ie. both builds actually use the libav fork I believe. It might be
nice to try and add a build against ffmpeg sometime.)

An odd looking "exclude" is needed as otherwise we end up with 3
builds - this is a workaround for:

https://github.com/travis-ci/travis-ci/issues/4681